Move class JobResourceTest to resources pkg
